pyproject.toml
[tool.poetry]
name = "backends"
version = "0.1.0"
description = ""
authors = ["dearsecret <dearsecret1217@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
djangorestframework = "^3.14.0"
markdown = "^3.4.4"
django-filter = "^23.2"
pylint-django = "^2.5.3"
django-environ = "^0.10.0"
django-channels = "^0.7.0"
channels-redis = "^4.1.0"
daphne = "^4.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
